Suggestion on build targets to match common software engineering concepts
Tutorial reads nice. It's nicely writen and straight to the concepts.
I enjoyed it but I got a bit confused by the terminology used in
the build targets section. I do get that emulation, in a software
context, is the use of an application program or device to imitate
the behavior of another program or device but given the low fidelity
(with regard hardware) obtained in the "software emulation" target
I believe that simulation (instead of emulation) would fit better,
since this is exactly its definition.

Looking at it theoretically, virtualization is a commonly used technique in CS allowing
developers to more easily test, reuse and ship systems.  When using virtualization,
there are often two approaches, emulation and  simulation. Emulation is generally
implemented through hardware-virtualization, capturing most aspects of the underlying
hardware in a virtual environment. It is typically using full-virtualization with type two
hypervisors (e.g.  Virtual-Box, VMWare Worksation or QEMU). Simulation is often
implemented with simpler abstractions such as state machines or through OS-virtualization,
using a shared kernel across both the host and guests. Simulation is widely considered the
most efficient virtualization method for its highest performance and fastest ”start-up”
time at the exponse of lower fidelity with regard the final hardware target.

This suggestion attempts to match these concepts with the most commonly
used ones in software engineering while at the same time, simplifies naming
for future reference/discussions.

Signed-off-by: Víctor Mayoral Vilches <v.mayoralv@gmail.com>